chromium: Use TUNE_ARCH in gn_arch_name()
It is a bit more generic than TARGET_ARCH and thus suits our purposes
better. It allows us to simplify gn_arch_name() as we do not need to
use regular expressions anymore.

Stop caring about MIPS for now, as TUNE_ARCH there can assume many
different values and it only makes sense to worry about it in the
future, once we've verified that everything is working as expected for
other architectures.