Update Chromium to 54.0.2840.59
This is the latest stable release at the time of writing. Moving to M54
allows us to drop all the patches we were carrying, as they have all
been integrated upstream in this milestone.

The custom Yocto toolchain GN entries have also been updated to match
the new format introduced in
https://codereview.chromium.org/2219953002 (see Chromium bug 634446).

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>