Improve cross-compilation support in the recipe
Chromium is a tricky project because, by default, it may end up
requiring 3 separate toolchains:
- GN's host_cpu and host_os variables deal with the system where the
  build is being performed (x86_64 on Linux, for example).
- GN's target_cpu and target_os variables deal with the system where
  Chromium will run (i586 or ARMv7a, for example).

However, the JIT code in V8 builds some binaries (mkpeephole and
mksnapshot as of M54) that need to be run during the build _with the
bit-width of the target CPU_. In other words, a x86_64 host targeting
ARMv7 needs to build 32-bit V8 JIT binaries and run them during the
build.

The solution upstream (which has their own clang toolchains) for the
scenario above is to use a x86_64 compiler targeting ARMv7 for most of
the build, and create the V8 binaries with -m32 and run them on the
host, assuming all the proper infrastructure to run 32-bit binaries on
64-bit Linux is already in place.

Doing so in Yocto is more complicated. Instead, we just use 2
toolchains, build the V8 binaries using the target toolchain and run
them on the host using QEMU, in a similar fashion to what the
recipes-gnome/gobject-introspection recipe does. V8 does need to be
patched to use our QEMU wrapper script though, and it is yet to be
determined whether this change is upstreamable in a more generalized
fashion.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>