Add more licenses to LIC_FILES_CHKSUM
Chromium itself is a 3-Clause-BSD-licensed project, but it is made up of
several other projects included in the tarball, many of which have
different licenses.

With the aid of Chromium's own tools/licenses.py module, list all other
copyright files for projects marked as being shipped as part of
Chromium (ie. not used only during the build or tests). The licenses
listed are a superset of the ones actually used though, as they include
projects used only by Android or iOS builds, for example.

None of those other licenses should change Chromium's main license.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>