Add a new option, "root-profile", for running Chromium as root
THIS SHOULD NOT BE USED IN PRODUCTION.

When this PACKAGECONFIG option is enabled, the wrapper script we use to
launch the Chromium binary is patches and we set the CHROME_USER_DATA_DIR
environment variable to "/root/chromium-profile", as by default Chromium
refused to run as root.

The idea here is to ease development and testing, which is why this option
is not enabled by default.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>