Improve the way debug flags are handled.
Despite commit 5aae1539c3 ("Remove debug flags from 32-bit builds"), recent
32-bit builds were still generating enormous (~5Gb) chrome binaries, even if
DEBUG_BUILD is not set. 32-bit binaries this size are not handled
correctly by binutils and file, which leads to all sorts of problems when
packaging Chromium.

This is mainly caused by the fact that when |is_official_build=true|, GN
sets |symbol_level=2|, which means -g2 will be passed to the compiler and a
binary will tons of debugging information will be generated. This is done
even though we used to remove ${DEBUG_FLAGS} (which includes -g by default)
from TARGET_CFLAGS on 32-bit architectures.

Fixing this requires adjusting both the recipe and GN:
1. Make sure -g is not passed on 32-bit architectures via DEBUG_FLAGS. -g is
   the same as -g2. -g1 generates an 800MB binary, which is a lot more
   manageable.
2. Explicitly pass symbol_level=0 to GN. This causes -g0 to be passed
   instead, so that if DEBUG_BUILD is not set GN will not create a huge
   debug binary anyway. Since our compiler flags are passed after GN's, -g0
   does not cause any issues if DEBUG_BUILD is set, as -g1 will be passed
   later.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>