Pass -g1 instead of -g for x86-64 as well.
Starting with Chromium 60, x86-64 is also having problems with -g being
passed to the compiler, although it is of a different nature than what
affects 32-bit platforms.

In this case, linking the chrome binary fails with several relocation
overflow errors such as:

    obj/third_party/WebKit/Source/core/loader/libloader.a(ModuleTreeLinker.o)(.debug_loc+0x1e9a5): error: relocation overflow: reference to local symbol 82 in obj/third_party/WebKit/Source/core/loader/libloader.a(ModuleTreeLinker.o)
    obj/third_party/WebKit/Source/core/libcore_generated.a(ScriptModule.o)(.debug_loc+0x253c): error: relocation overflow: reference to local symbol 31 in obj/third_party/WebKit/Source/core/libcore_generated.a(ScriptModule.o)

which can also be reproduced on a regular desktop build when system GCC and
system binutils (tested with Fedora 26's GCC 7 and binutils 2.27).

Fixes #3.