Explicitly set the use_gold GN variable.
So far we were assuming ld.gold was always present and preferred. While this
tends not to be an issue and upstream only tests gold in its builds (it's
faster and uses less memory), we need to respect the ld-is-gold
DISTRO_FEATURE.

Doing so also helps when building the native bits, as gold's presence in the
host system is less likely than in the target toolchain.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>